requirements.txt
camelot-py[cv]>=0.10.1
markitdown==0.1.2
openai>=1.30.1
pdf2image>=1.17.0
pdfplumber>=0.11.6
Pillow>=11.0.0
pymupdf>=1.23.7
python-dotenv>=1.0.1
pytesseract>=0.3.10
streamlit>=1.25.0
tabulate>=0.9.0
stqdm>=0.0.5
supabase>=2.0.0